% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCDF.R
\name{CCDF}
\alias{CCDF}
\title{Function to compute (un)conditional cumulative distribution function (CDF), used by plot_CCDF function.}
\usage{
CCDF(
  Y,
  X,
  Z = NULL,
  method = c("linear regression", "logistic regression", "RF"),
  fast = TRUE,
  space_y = FALSE,
  number_y = length(Y)
)
}
\arguments{
\item{Y}{a numeric vector of size \code{n} containing the
preprocessed expressions from \code{n} samples (or cells).}

\item{X}{a data frame containing numeric or factor vector(s) of size \code{n}
containing the variable(s) to be tested (the condition(s) to be tested).}

\item{Z}{a data frame containing numeric or factor vector(s) of size \code{n}
containing the covariate(s).}

\item{method}{a character string indicating which method to use to
compute the CCDF, either \code{'linear regression'}, \code{'logistic regression'}
and  \code{'permutations'} or \code{'RF'} for Random Forests.
Default is \code{'linear regression'} since it is the method used in the test.}

\item{fast}{a logical flag indicating whether the fast implementation of
logistic regression should be used. Only if \code{'dist_permutations'} is specified.
Default is \code{TRUE}.}

\item{space_y}{a logical flag indicating whether the y thresholds are spaced. 
When \code{space_y} is \code{TRUE}, a regular sequence between the minimum and 
the maximum of the observations is used. Default is \code{FALSE}.}

\item{number_y}{an integer value indicating the number of y thresholds (and therefore
the number of regressions) to perform the test. Default is \code{length(Y)}.}
}
\description{
Function to compute (un)conditional cumulative distribution function (CDF), used by plot_CCDF function.
}
