% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF_permut.R
\name{CDF_permut}
\alias{CDF_permut}
\title{CDF_permut function}
\usage{
CDF_permut(
  Y,
  X,
  Z = NULL,
  method = "logistic regression",
  fast = TRUE,
  coef = CDF$coeff,
  list_boot = list_boot,
  list_OOB = list_OOB
)
}
\arguments{
\item{X}{}
}
\description{
CDF_permut function
}
