% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{permut}
\alias{permut}
\title{permut}
\usage{
permut(
  Y,
  X,
  Z = NULL,
  distance = "L2",
  n_perm,
  method = "logistic regression",
  parallel = TRUE,
  n_cpus = NULL,
  fast = TRUE
)
}
\arguments{
\item{Y}{}
}
\description{
permut
}
\examples{
Y <- c(rep(0,40),rnorm(n = 160))
X <- as.factor(rbinom(n=200, size = 1, prob = 0.5))
Z <- rnorm(n = 100)
permut(Y,X,n_perm=100)
}
