% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.citcdf.R
\name{plot.citcdf}
\alias{plot.citcdf}
\title{Function for plotting the CDF of all the genes within a gene set or for one gene}
\usage{
\method{plot}{citcdf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{cit_gsa}}}

\item{...}{further arguments to be passed}
}
\value{
a \code{ggplot2} of gene-wise or geneset ccdf
}
\description{
Function for plotting the CDF of all the genes within a gene set or for one gene
}
\examples{
#TO DO
}
