% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccdf_cit_gsa.R
\name{plot_ccdf_cit_gsa}
\alias{plot_ccdf_cit_gsa}
\title{Function for plotting the CDF of all the genes within a gene set}
\usage{
plot_ccdf_cit_gsa(ccdf, number_y = length(ccdf[[1]]$y))
}
\arguments{
\item{ccdf}{an R object that come from the CCDF function of the package}

\item{number_y}{an integer value indicating the number of y thresholds for the summary curves. Default is \code{length(Y)}.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Function for plotting the CDF of all the genes within a gene set
}
\examples{
#TO DO
}
