% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvals.R
\name{plot_pvals}
\alias{plot_pvals}
\title{Plot of gene-wise p-values}
\usage{
plot_pvals(pvals, ...)
}
\arguments{
\item{pvals}{a vector of length \code{n} containing the raw p-values for
each gene}
}
\value{
a plot of sorted gene-wise p-values
}
\description{
This function prints the sorted exact p-values along with the Benjamini-Hochberg
limit and the 5% threshold
}
