% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKAT_davies.R
\name{SKAT_davies}
\alias{SKAT_davies}
\title{Functions from CompQuadForm package}
\usage{
SKAT_davies(
  q,
  lambda,
  h = rep(1, length(lambda)),
  delta = rep(0, length(lambda)),
  sigma = 0,
  lim = 10000,
  acc = 1e-04
)
}
\arguments{
\item{X}{}
}
\description{
Functions from CompQuadForm package
}
