% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCDF.R
\name{CCDF}
\alias{CCDF}
\title{CDF function}
\usage{
CCDF(
  Y,
  X,
  Z = NULL,
  method = "linear regression",
  fast = TRUE,
  space_y = FALSE,
  number_y = length(Y)
)
}
\arguments{
\item{X}{}
}
\description{
CDF function
}
