% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdf_testing.R
\name{ccdf_testing}
\alias{ccdf_testing}
\title{CDF function}
\usage{
ccdf_testing(
  exprmat = NULL,
  variables2test = NULL,
  covariates = NULL,
  method = "logistic regression",
  norm = "L2",
  test = c("permutations", "asymptotic"),
  n_perm = 100,
  parallel = TRUE,
  n_cpus = 15,
  fast = TRUE
)
}
\arguments{
\item{exprmat}{}
}
\description{
CDF function
}
\examples{
X <- as.factor(rbinom(n=50, size = 1, prob = 0.5))
Y <- matrix(rnorm(n = 1000,0,1),nrow=10,ncol=10)
Y <- ((X==1)*rnorm(n = 50,0,1)) + ((X==0)*rnorm(n = 50,2,1))
Y <- replicate(10,Y)
Z <- rnorm(n = 100)
profvis(res <- ccdf_testing(t(Y),X,method="logistic regression",n_cpus=15))

}
