% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit_perm.R
\name{cit_perm}
\alias{cit_perm}
\title{Permutation test for conditional independence}
\usage{
cit_perm(Y, X, Z = NULL, n_perm = 100, space_y = FALSE, number_y = length(Y))
}
\arguments{
\item{Y}{a numeric vector of length \code{n} to test for conditional independence 
with \code{X} adjusted on \code{Z}}

\item{X}{a data frame of size \code{n x p} of numeric or factor vector(s) 
containing the variable(s) to be tested for conditional independence 
against \code{X} adjusted on \code{Z}. Multiple variables are not supported 
for permutation.}

\item{Z}{a data.frame of size \code{n x 1} of numeric or factor vector 
containing the covariate to condition the independence 
test upon. Multiple covariates are not supported for permutation.}

\item{n_perm}{the number of permutations. Default is \code{100}.}

\item{space_y}{a logical flag indicating whether the y thresholds are spaced out. 
When \code{space_y} is \code{TRUE}, a regular sequence between the minimum and 
the maximum of the observations is used. Default is \code{FALSE}.}

\item{number_y}{an integer value indicating the number of y thresholds (and therefore
the number of regressions) to perform the test. Default is \code{length(Y)}.}
}
\value{
A data frame with the following elements:
\itemize{
  \item \code{score} contains the test statistic for a given gene.
  \item \code{raw_pval} contains the raw p-values for a given gene computed from \code{n_perm} permutations.
}
}
\description{
Permutation test for conditional independence
}
\examples{

if(interactive()){
X <- as.factor(rbinom(n=100, size = 1, prob = 0.5))
Y <- ((X==1)*rnorm(n = 50,0,1)) + ((X==0)*rnorm(n = 50,0.5,1))
res_perm <- cit_perm(Y,data.frame(X=X),n_perm=10)}
}
