% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.R
\name{permut}
\alias{permut}
\title{permut}
\usage{
permut(
  Y,
  X,
  Z = NULL,
  norm = "L2",
  n_perm = 100,
  method = "logistic regression",
  parallel = TRUE,
  n_cpus = 15,
  fast = TRUE
)
}
\arguments{
\item{Y}{}
}
\description{
permut
}
\examples{
Y <- rnorm(n = 200)
X <- as.factor(rbinom(n=200, size = 1, prob = 0.5))
Z <- rnorm(n = 100)
profvis(permut(Y,X))
}
