% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cit_multi.R
\name{plot.cit_multi}
\alias{plot.cit_multi}
\title{Plot of gene-wise p-values}
\usage{
\method{plot}{cit_multi}(x, ..., nominal_level = 0.05)
}
\arguments{
\item{x}{an object of class \code{\link{cit_multi}}.}

\item{...}{further arguments to be passed}

\item{nominal_level}{a nominal testing level between 0 and 1. 
Default is 5\%: \code{0.05}.}
}
\value{
a \code{ggplot2} of sorted gene-wise p-values
}
\description{
Plotting the sorted exact p-values along with the Benjamini-Hochberg
limit and the nominal threshold
}
\examples{
n <- 100
p <- 500
X1 <- as.factor(rbinom(n=n, size = 1, prob = 0.5))
Y <- replicate(p, ((X1==1)*rnorm(n = n,0,1)) + ((X1==0)*rnorm(n = n, 0.5, 1)))

res_asymp <- cit_multi(M=data.frame(Y=Y),
                      X=data.frame(X = X1), 
                      test="asymptotic",
                      parallel=FALSE)
plot(res_asymp)

}
