% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvals.R
\name{plot_pvals}
\alias{plot_pvals}
\title{Plot of gene-wise p-values}
\usage{
plot_pvals(pvals, nominal_level = 0.05)
}
\arguments{
\item{pvals}{a vector of length \code{n} containing the raw p-values for
each gene}

\item{nominal_level}{a nominal testing level between 0 and 1. 
Default is 5\%: \code{0.05}.}
}
\value{
a \code{ggplot2} of sorted gene-wise p-values
}
\description{
Plotting the sorted exact p-values along with the Benjamini-Hochberg
limit and the nominal threshold
}
\examples{
plot_pvals(runif(100,0,1))
}
