% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_asymp.R
\name{test_asymp}
\alias{test_asymp}
\title{Asymptotic test}
\usage{
test_asymp(
  Y,
  X,
  Z = NULL,
  space_y = FALSE,
  number_y = length(unique(Y)),
  log = FALSE
)
}
\arguments{
\item{X}{}
}
\description{
Asymptotic test
}
