% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_perm.R
\name{test_perm}
\alias{test_perm}
\title{Permutation test using the test statistic}
\usage{
test_perm(Y, X, Z = NULL, n_perm = 100, parallel = TRUE, n_cpus = NULL)
}
\arguments{
\item{n_cpus}{}
}
\description{
Permutation test using the test statistic
}
