% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_perm.R
\name{test_perm}
\alias{test_perm}
\title{Permutation test using the test statistic}
\usage{
test_perm(
  Y,
  X,
  Z = NULL,
  n_perm = 100,
  parallel = FALSE,
  n_cpus = NULL,
  space_y = FALSE,
  number_y = 5,
  log = FALSE,
  keep_zeros = TRUE
)
}
\arguments{
\item{space_y}{a logical flag indicating whether the y thresholds are spaced. 
When \code{space_y} is \code{TRUE}, a regular sequence between the minimum and 
the maximum of the observations is used. Default is \code{FALSE}.}

\item{number_y}{a integer indicating the number of y thresholds (and therefore
the number of regressions) to perform the test. Default is \code{NULL}.}
}
\description{
Permutation test using the test statistic
}
