% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_ccdf.R
\name{weights_ccdf}
\alias{weights_ccdf}
\title{Computation of weights for CCDF, used in permut function when \code{dist_permuations} is specified.}
\usage{
weights_ccdf(Y, X, Z = NULL)
}
\arguments{
\item{Y}{a numeric vector of size \code{n} containing the
preprocessed expressions from \code{n} samples (or cells).}

\item{X}{a numeric or factor vector(s) of size \code{n}
containing the variable to be tested (the condition to be tested). 
Multiple variables are not allowed}

\item{Z}{a numeric vector of size \code{n}
containing the covariate.
Multiple variables are not allowed}
}
\description{
Computation of weights for CCDF, used in permut function when \code{dist_permuations} is specified.
}
\keyword{internal}
